refactor(notes): extract note-editor markdown component map
note-editor.tsx was 1176 lines, ~240 of them the ReactMarkdown components
override map (read-mode preview: links, wikilinks, callouts, math, checkboxes,
tables, footnotes, images, headings) plus the renderCellWithCheckboxes helper.

Move both into note-markdown-components.tsx as useNoteMarkdownComponents(),
taking { toggleCheckbox, notesRef, previewScrollRef }. The referential-stability
contract is preserved verbatim: the same useMemo([toggleCheckbox]) memo and the
notesRef/previewScrollRef ref-closure pattern, so the map stays stable across
unrelated store updates (no preview re-parse / KaTeX remount). No behaviour or
styling change — component bodies + className strings copied verbatim.

Host note-editor.tsx drops 1176 -> 903 and sheds 6 now-unused imports; net
eslint warnings in the file went 4 -> 2 (the 2 remaining are pre-existing).

Verified: type-check clean; eslint clean (warnings-only, reduced); full suite
(1128) green.
