+++ changelogs/v2.4.14.md
@@ -11,3 +11,4 @@
 - Also split the codebase semantic-indexing queries (files, symbols, relations, and the Architecture/Module-Map rollups) into their own file, the same way. Purely internal; no behaviour change.
 - Split the Tools settings panel (MCP servers + custom HTTP services) — its header editor, connection tester, OAuth sign-in control, add/edit forms, list row, and per-server tool checklist now live in their own files, sharing one helpers module. Purely internal; the settings screen looks and behaves exactly the same.
 - Split the Agent's Git panel — its file-section, changed-file row, inline-diff renderer, and git status-mapping helpers now live in their own files. Purely internal; the Git view looks and behaves exactly the same.
+- Split the note editor's read-mode preview renderer (the ReactMarkdown component overrides — links, wikilinks, callouts, math, checkboxes, tables, footnotes, etc.) into its own module. Purely internal, with the same referential-stability guarantees preserved; the rendered note looks and behaves exactly the same.
